Rename albumService to album
